+++ demo/views/chrome/footer/footer.js
@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import I18n from 'i18n-js';
 
 // Carbon