+++ src/components/i18n/i18n.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import i18n from 'i18n-js';
 import _marked from 'marked';
 import { tagComponent } from '../../utils/helpers/tags';
@@ -33,7 +34,7 @@ class I18n extends React.Component {
      * @type {Boolean}
      * @default false
      */
-    markdown: React.PropTypes.bool,
+    markdown: PropTypes.bool,
 
     /**
      * Whether to enclose the text in a <span> or a <div>
@@ -42,7 +43,7 @@ class I18n extends React.Component {
      * @type {Boolean}
      * @default true
      */
-    inline: React.PropTypes.bool,
+    inline: PropTypes.bool,
 
     /**
      * The key to lookup for a localised value
@@ -51,7 +52,7 @@ class I18n extends React.Component {
      * @type {String}
      * @default undefined
      */
-    scope: React.PropTypes.string,
+    scope: PropTypes.string,
 
     /**
      * Additional options to pass to I18n
@@ -60,7 +61,7 @@ class I18n extends React.Component {
      * @type {Object}
      * @default undefined
      */
-    options: React.PropTypes.object
+    options: PropTypes.object
   }
 
   static defaultProps = {