+++ src/components/inline-inputs/__spec__.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import TestUtils from 'react/lib/ReactTestUtils';
+import TestUtils from 'react-dom/test-utils';
 import classNames from 'classnames';
 import Row from './../row';
 import Textbox from './../textbox';