+++ src/components/simple-color-picker/color-option/color-option.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 
 import Input from './../../../utils/decorators/input';
 import Icon from './../../icon';
@@ -20,7 +21,7 @@ class ColorOption extends React.Component {
      * @property color
      * @type {String}
      */
-    color: React.PropTypes.string,
+    color: PropTypes.string,
 
     /**
      * the input name.
@@ -28,7 +29,7 @@ class ColorOption extends React.Component {
      * @property name
      * @type {String}
      */
-    name: React.PropTypes.string,
+    name: PropTypes.string,
 
     /**
      * called when the user selects or deselects this color option.
@@ -36,7 +37,7 @@ class ColorOption extends React.Component {
      * @property onChange
      * @type {Function}
      */
-    onChange: React.PropTypes.func,
+    onChange: PropTypes.func,
 
     /**
      * determines if this color option is selected or unselected.
@@ -44,7 +45,7 @@ class ColorOption extends React.Component {
      * @property checked
      * @type {Boolean}
      */
-    checked: React.PropTypes.bool
+    checked: PropTypes.bool
   }
 
   /**