+++ src/components/table/table-cell/table-cell.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import { validProps } from '../../../utils/ether';
 import { tagComponent } from '../../../utils/helpers/tags';
@@ -28,7 +29,7 @@ class TableCell extends React.Component {
      * @property align
      * @type {String}
      */
-    align: React.PropTypes.string,
+    align: PropTypes.string,
 
     /**
      * Defines the cell type to be an action - used for the delete cell.
@@ -36,7 +37,7 @@ class TableCell extends React.Component {
      * @property action
      * @type {Boolean}
      */
-    action: React.PropTypes.bool
+    action: PropTypes.bool
   }
 
   /**