+++ src/components/table/table-subheader/__spec__.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import TestUtils from 'react/lib/ReactTestUtils';
+import TestUtils from 'react-dom/test-utils';
 import { Table, TableRow } from './../table';
 import TableSubheader from './../table-subheader';
 import { shallow } from 'enzyme';